+++ app/views/pages/index.ejs
@@ -201,7 +201,7 @@
 		for (var i=0;i<links.length;i++){
 		var link = links[i];				
 		$.ajax({
-		url: '/getBlogData',
+		url: '/b/getBlogData',
 		type: 'POST',
 		dataType: 'json',
 		timeout: 4000,
